requirements.txt
openai==1.3.0
python-dotenv==1.0.0
colorama==0.4.6
requests>=2.25.0
flask>=2.0.0
flask-cors>=3.0.0
